+++ jabgui/src/main/java/org/jabref/gui/actions/StandardActions.java
@@ -135,9 +135,10 @@ public enum StandardActions implements Action {
     EXPAND_ALL(Localization.lang("Expand all")),
     COLLAPSE_ALL(Localization.lang("Collapse all")),
 
-    NEW_ENTRY(Localization.lang("New entry"), IconTheme.JabRefIcons.ADD_ENTRY, KeyBinding.NEW_ENTRY),
-    NEW_ARTICLE(Localization.lang("New article"), IconTheme.JabRefIcons.ADD_ARTICLE),
-    NEW_ENTRY_FROM_PLAIN_TEXT(Localization.lang("New entry from plain text"), IconTheme.JabRefIcons.NEW_ENTRY_FROM_PLAIN_TEXT),
+    NEW_INSTANT_ENTRY(Localization.lang("Add empty entry"), IconTheme.JabRefIcons.ADD_ARTICLE),
+    NEW_ENTRY_UNIFIED(Localization.lang("Create new entry"), IconTheme.JabRefIcons.ADD_ENTRY, KeyBinding.NEW_ENTRY),
+    NEW_ENTRY_PLAINTEXT(Localization.lang("Interpret citations"), IconTheme.JabRefIcons.NEW_ENTRY_PLAINTEXT, KeyBinding.NEW_ENTRY_PLAINTEXT),
+
     LIBRARY_PROPERTIES(Localization.lang("Library properties")),
     FIND_DUPLICATES(Localization.lang("Find duplicates"), IconTheme.JabRefIcons.FIND_DUPLICATES),
     MERGE_ENTRIES(Localization.lang("Merge entries"), IconTheme.JabRefIcons.MERGE_ENTRIES, KeyBinding.MERGE_ENTRIES),