+++ jabgui/src/main/java/org/jabref/gui/keyboard/KeyBinding.java
@@ -69,10 +69,11 @@ public enum KeyBinding {
     IMPORT_INTO_NEW_DATABASE("Import into new library", Localization.lang("Import into new library"), "ctrl+alt+I", KeyBindingCategory.FILE),
     MERGE_ENTRIES("Merge entries", Localization.lang("Merge entries"), "ctrl+M", KeyBindingCategory.TOOLS),
 
+    // :MYTODO: Keybindings for New Entry Unified. Make extra keybinding to keep the plaintext option.
     NEW_ARTICLE("New article", Localization.lang("New article"), "ctrl+shift+A", KeyBindingCategory.BIBTEX),
     NEW_BOOK("New book", Localization.lang("New book"), "ctrl+shift+B", KeyBindingCategory.BIBTEX),
     NEW_ENTRY("New entry", Localization.lang("New entry"), "ctrl+N", KeyBindingCategory.BIBTEX),
-    NEW_ENTRY_FROM_PLAIN_TEXT_ONLINE("New entry from plain text", Localization.lang("New entry from plain text (online)"), "ctrl+shift+N", KeyBindingCategory.BIBTEX),
+    NEW_ENTRY_PLAINTEXT("New entry from citation", Localization.lang("New entry from plain text (online)"), "ctrl+shift+N", KeyBindingCategory.BIBTEX),
     NEW_INBOOK("New inbook", Localization.lang("New inbook"), "ctrl+shift+I", KeyBindingCategory.BIBTEX),
     NEW_MASTERSTHESIS("New mastersthesis", Localization.lang("New mastersthesis"), "ctrl+shift+M", KeyBindingCategory.BIBTEX),
     NEW_PHDTHESIS("New phdthesis", Localization.lang("New phdthesis"), "ctrl+shift+T", KeyBindingCategory.BIBTEX),