+++ jabgui/src/main/java/org/jabref/gui/newentryunified/NewEntryUnifiedApproach.java
@@ -1,0 +1,8 @@
+package org.jabref.gui.newentryunified;
+
+public enum NewEntryUnifiedApproach {
+    CREATE_ENTRY,
+    LOOKUP_IDENTIFIER,
+    INTERPRET_CITATIONS,
+    SPECIFY_BIBTEX,
+}