+++ jabgui/src/main/java/org/jabref/gui/newentryunified/NewEntryUnifiedView.java
@@ -1,0 +1,508 @@
+package org.jabref.gui.newentryunified;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
+
+import javafx.application.Platform;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
+import javafx.fxml.FXML;
+import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
+import javafx.scene.control.ComboBox;
+import javafx.scene.control.Label;
+import javafx.scene.control.RadioButton;
+import javafx.scene.control.Tab;
+import javafx.scene.control.TabPane;
+import javafx.scene.control.TextArea;
+import javafx.scene.control.TextField;
+import javafx.scene.control.TitledPane;
+import javafx.scene.control.ToggleGroup;
+import javafx.scene.control.Tooltip;
+import javafx.scene.layout.FlowPane;
+import javafx.stage.Screen;
+import javafx.stage.Stage;
+
+import org.jabref.gui.ClipBoardManager;
+import org.jabref.gui.DialogService;
+import org.jabref.gui.LibraryTab;
+import org.jabref.gui.StateManager;
+import org.jabref.gui.preferences.GuiPreferences;
+import org.jabref.gui.search.SearchType;
+import org.jabref.gui.util.BaseDialog;
+import org.jabref.gui.util.ControlHelper;
+import org.jabref.gui.util.IconValidationDecorator;
+import org.jabref.gui.util.UiTaskExecutor;
+import org.jabref.gui.util.ViewModelListCellFactory;
+import org.jabref.logic.ai.AiService;
+import org.jabref.logic.importer.IdBasedFetcher;
+import org.jabref.logic.importer.WebFetcher;
+import org.jabref.logic.importer.fetcher.DoiFetcher;
+import org.jabref.logic.importer.plaincitation.PlainCitationParserChoice;
+import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.util.TaskExecutor;
+import org.jabref.model.database.BibDatabaseMode;
+import org.jabref.model.entry.BibEntry;
+import org.jabref.model.entry.BibEntryType;
+import org.jabref.model.entry.BibEntryTypesManager;
+import org.jabref.model.entry.types.BiblatexAPAEntryTypeDefinitions;
+import org.jabref.model.entry.types.BiblatexEntryTypeDefinitions;
+import org.jabref.model.entry.types.BiblatexSoftwareEntryTypeDefinitions;
+import org.jabref.model.entry.types.BibtexEntryTypeDefinitions;
+import org.jabref.model.entry.types.EntryType;
+import org.jabref.model.entry.types.IEEETranEntryTypeDefinitions;
+import org.jabref.model.entry.types.StandardEntryType;
+import org.jabref.model.strings.StringUtil;
+import org.jabref.model.util.FileUpdateMonitor;
+
+import com.airhacks.afterburner.injection.Injector;
+import com.airhacks.afterburner.views.ViewLoader;
+import com.tobiasdiez.easybind.EasyBind;
+import de.saxsys.mvvmfx.utils.validation.visualization.ControlsFxVisualizer;
+import jakarta.inject.Inject;
+
+public class NewEntryUnifiedView extends BaseDialog<BibEntry> {
+    private NewEntryUnifiedViewModel viewModel;
+
+    private final NewEntryUnifiedApproach initialApproach;
+    private NewEntryUnifiedApproach currentApproach;
+
+    private final GuiPreferences guiPreferences;
+    private final NewEntryUnifiedPreferences preferences;
+    private final LibraryTab libraryTab;
+    private final DialogService dialogService;
+    @Inject private StateManager stateManager;
+    @Inject private TaskExecutor taskExecutor;
+    @Inject private AiService aiService;
+    @Inject private FileUpdateMonitor fileUpdateMonitor;
+
+    private final ControlsFxVisualizer visualizer;
+
+    @FXML private ButtonType generateButtonType;
+    private Button generateButton;
+
+    @FXML private TabPane tabs;
+    @FXML private Tab tabCreateEntry;
+    @FXML private Tab tabLookupIdentifier;
+    @FXML private Tab tabInterpretCitations;
+    @FXML private Tab tabSpecifyBibtex;
+
+    @FXML private TitledPane entryRecommendedTitle;
+    @FXML private FlowPane entryRecommended;
+    @FXML private TitledPane entryOtherTitle;
+    @FXML private FlowPane entryOther;
+    @FXML private TitledPane entryCustomTitle;
+    @FXML private FlowPane entryCustom;
+
+    @FXML private TextField idText;
+    @FXML private RadioButton idLookupGuess;
+    @FXML private RadioButton idLookupSpecify;
+    @FXML private ComboBox<IdBasedFetcher> idFetcher;
+    @FXML private Label idErrorInvalidText;
+    @FXML private Label idErrorInvalidFetcher;
+
+    @FXML private TextArea interpretText;
+    @FXML private ComboBox<PlainCitationParserChoice> interpretParser;
+
+    @FXML private TextArea bibtexText;
+
+    private BibEntry result;
+
+    public NewEntryUnifiedView(NewEntryUnifiedApproach initialApproach, GuiPreferences preferences, LibraryTab libraryTab, DialogService dialogService) {
+        this.initialApproach = initialApproach;
+        this.currentApproach = initialApproach;
+
+        this.guiPreferences = preferences;
+        this.preferences = preferences.getNewEntryUnifiedPreferences();
+        this.libraryTab = libraryTab;
+        this.dialogService = dialogService;
+
+        visualizer = new ControlsFxVisualizer();
+        this.setTitle(Localization.lang("New Entry"));
+        ViewLoader.view(this).load().setAsDialogPane(this);
+
+        generateButton = (Button) this.getDialogPane().lookupButton(generateButtonType);
+        generateButton.getStyleClass().add("customGenerateButton");
+
+        ((Stage) getDialogPane().getScene().getWindow()).setMinWidth(400);
+        ((Stage) getDialogPane().getScene().getWindow()).setMinHeight(300);
+
+        ControlHelper.setAction(generateButtonType, getDialogPane(), event -> execute());
+        setOnCloseRequest(e -> cancel());
+        setResultConverter(button -> {
+            return result;
+        });
+
+        getDialogPane().disableProperty().bind(viewModel.executingProperty());
+
+        finalizeTabs();
+        tabs.requestFocus();
+    }
+
+    private void finalizeTabs() {
+        NewEntryUnifiedApproach approach = initialApproach;
+        if (approach == null) {
+            approach = preferences.getLatestApproach();
+        }
+
+        switch (approach) {
+            case NewEntryUnifiedApproach.CREATE_ENTRY:
+                tabs.getSelectionModel().select(tabCreateEntry);
+                switchCreateEntry();
+                break;
+            case NewEntryUnifiedApproach.LOOKUP_IDENTIFIER:
+                tabs.getSelectionModel().select(tabLookupIdentifier);
+                switchLookupIdentifier();
+                break;
+            case NewEntryUnifiedApproach.INTERPRET_CITATIONS:
+                tabs.getSelectionModel().select(tabInterpretCitations);
+                switchInterpretCitations();
+                break;
+            case NewEntryUnifiedApproach.SPECIFY_BIBTEX:
+                tabs.getSelectionModel().select(tabSpecifyBibtex);
+                switchSpecifyBibtex();
+                break;
+        }
+
+        tabCreateEntry.setOnSelectionChanged(event -> switchCreateEntry());
+        tabLookupIdentifier.setOnSelectionChanged(event -> switchLookupIdentifier());
+        tabInterpretCitations.setOnSelectionChanged(event -> switchInterpretCitations());
+        tabSpecifyBibtex.setOnSelectionChanged(event -> switchSpecifyBibtex());
+    }
+
+    @FXML
+    public void initialize() {
+        viewModel = new NewEntryUnifiedViewModel(guiPreferences, libraryTab, dialogService, stateManager, (UiTaskExecutor) taskExecutor, aiService, fileUpdateMonitor);
+
+        visualizer.setDecoration(new IconValidationDecorator());
+
+        EasyBind.subscribe(
+            viewModel.executedSuccessfullyProperty(),
+            succeeded -> {
+                if (succeeded) {
+                    onSuccessfulExecution();
+                }
+            });
+
+        initializeCreateEntry();
+        initializeLookupIdentifier();
+        initializeInterpretCitations();
+        initializeSpecifyBibTex();
+    }
+
+    private void initializeCreateEntry() {
+        entryRecommendedTitle.managedProperty().bind(entryRecommendedTitle.visibleProperty());
+        entryRecommended.managedProperty().bind(entryRecommended.visibleProperty());
+        entryOtherTitle.managedProperty().bind(entryOtherTitle.visibleProperty());
+        entryOther.managedProperty().bind(entryOther.visibleProperty());
+        entryCustomTitle.managedProperty().bind(entryCustomTitle.visibleProperty());
+        entryCustom.managedProperty().bind(entryCustom.visibleProperty());
+
+        final boolean isBiblatexMode = libraryTab.getBibDatabaseContext().isBiblatexMode();
+
+        List<BibEntryType> recommendedEntries;
+        List<BibEntryType> otherEntries;
+        if (isBiblatexMode) {
+            recommendedEntries = BiblatexEntryTypeDefinitions.RECOMMENDED;
+            otherEntries = new ArrayList<>(BiblatexEntryTypeDefinitions.ALL);
+            otherEntries.removeAll(recommendedEntries);
+            otherEntries.addAll(BiblatexSoftwareEntryTypeDefinitions.ALL);
+            otherEntries.addAll(BiblatexAPAEntryTypeDefinitions.ALL);
+        } else {
+            recommendedEntries = BibtexEntryTypeDefinitions.RECOMMENDED;
+            otherEntries = new ArrayList<>(BiblatexEntryTypeDefinitions.ALL);
+            otherEntries.removeAll(recommendedEntries);
+            otherEntries.addAll(IEEETranEntryTypeDefinitions.ALL);
+        }
+        addEntriesToPane(entryRecommended, recommendedEntries);
+        addEntriesToPane(entryOther, otherEntries);
+
+        final BibEntryTypesManager entryTypesManager = Injector.instantiateModelOrService(BibEntryTypesManager.class);
+        final BibDatabaseMode customTypesDatabaseMode = isBiblatexMode ? BibDatabaseMode.BIBLATEX : BibDatabaseMode.BIBTEX;
+        final List<BibEntryType> customEntries = entryTypesManager.getAllCustomTypes(customTypesDatabaseMode);
+        if (customEntries.isEmpty()) {
+            entryCustomTitle.setVisible(false);
+        } else {
+            addEntriesToPane(entryCustom, customEntries);
+        }
+    }
+
+    private void initializeLookupIdentifier() {
+        idText.setPromptText(Localization.lang("Enter the reference identifier to search for."));
+        idText.textProperty().bindBidirectional(viewModel.idTextProperty());
+        final String clipboardText = ClipBoardManager.getContents().trim();
+        if (!StringUtil.isBlank(clipboardText) && !clipboardText.contains("\n")) {
+            idText.setText(clipboardText);
+            idText.selectAll();
+        }
+
+        idLookupGuess.selectedProperty().addListener(
+            new ChangeListener<Boolean>() {
+                public void changed(ObservableValue<? extends Boolean> observable, Boolean oldValue, Boolean newValue) {
+                    preferences.setIdLookupGuessing(newValue);
+                }
+            });
+
+        ToggleGroup toggleGroup = new ToggleGroup();
+        idLookupGuess.setToggleGroup(toggleGroup);
+        idLookupSpecify.setToggleGroup(toggleGroup);
+
+        if (preferences.getIdLookupGuessing()) {
+            idLookupGuess.selectedProperty().set(true);
+        } else {
+            idLookupSpecify.selectedProperty().set(true);
+        }
+
+        idFetcher.itemsProperty().bind(viewModel.idFetchersProperty());
+        new ViewModelListCellFactory<IdBasedFetcher>().withText(WebFetcher::getName).install(idFetcher);
+        idFetcher.disableProperty().bind(idLookupSpecify.selectedProperty().not());
+        idFetcher.valueProperty().bindBidirectional(viewModel.idFetcherProperty());
+        IdBasedFetcher initialFetcher = fetcherFromName(preferences.getLatestIdFetcher(), idFetcher.getItems());
+        if (initialFetcher == null) {
+            final IdBasedFetcher defaultFetcher = new DoiFetcher(guiPreferences.getImportFormatPreferences());
+            initialFetcher = fetcherFromName(defaultFetcher.getName(), idFetcher.getItems());
+        }
+        idFetcher.setValue(initialFetcher);
+        idFetcher.setOnAction(event -> {
+            preferences.setLatestIdFetcher(idFetcher.getValue().getName());
+        });
+
+        idErrorInvalidText.visibleProperty().bind(viewModel.idTextValidatorProperty().not());
+        idErrorInvalidFetcher.visibleProperty().bind(idLookupSpecify.selectedProperty().and(viewModel.idFetcherValidatorProperty().not()));
+    }
+
+    private void initializeInterpretCitations() {
+        interpretText.setPromptText(Localization.lang("Enter the plain citations to parse, separated by blank lines."));
+        interpretText.textProperty().bindBidirectional(viewModel.interpretTextProperty());
+        final String clipboardText = ClipBoardManager.getContents().trim();
+        if (!StringUtil.isBlank(clipboardText)) {
+            interpretText.setText(clipboardText);
+            interpretText.selectAll();
+        }
+
+        interpretParser.itemsProperty().bind(viewModel.interpretParsersProperty());
+        new ViewModelListCellFactory<PlainCitationParserChoice>().withText(PlainCitationParserChoice::getLocalizedName).install(interpretParser);
+        interpretParser.valueProperty().bindBidirectional(viewModel.interpretParserProperty());
+        PlainCitationParserChoice initialParser = parserFromName(preferences.getLatestInterpretParser(), interpretParser.getItems());
+        if (initialParser == null) {
+            final PlainCitationParserChoice defaultParser = PlainCitationParserChoice.RULE_BASED;
+            initialParser = parserFromName(defaultParser.getLocalizedName(), interpretParser.getItems());
+        }
+        interpretParser.setValue(initialParser);
+        interpretParser.setOnAction(event -> {
+            preferences.setLatestInterpretParser(interpretParser.getValue().getLocalizedName());
+        });
+    }
+
+    private void initializeSpecifyBibTex() {
+        bibtexText.setPromptText(Localization.lang("Enter the Bib(La)TeX sources to generate entries from."));
+        bibtexText.textProperty().bindBidirectional(viewModel.bibtexTextProperty());
+        final String clipboardText = ClipBoardManager.getContents().trim();
+        if (!StringUtil.isBlank(clipboardText)) {
+            bibtexText.setText(clipboardText);
+            bibtexText.selectAll();
+        }
+    }
+
+    @FXML
+    private void switchCreateEntry() {
+        if (!tabCreateEntry.isSelected()) {
+            return;
+        }
+
+        currentApproach = NewEntryUnifiedApproach.CREATE_ENTRY;
+        preferences.setLatestApproach(NewEntryUnifiedApproach.CREATE_ENTRY);
+
+        if (generateButton != null) {
+            generateButton.disableProperty().unbind();
+            generateButton.setDisable(true);
+            generateButton.setText("Select");
+        }
+    }
+
+    @FXML
+    private void switchLookupIdentifier() {
+        if (!tabLookupIdentifier.isSelected()) {
+            return;
+        }
+
+        currentApproach = NewEntryUnifiedApproach.LOOKUP_IDENTIFIER;
+        preferences.setLatestApproach(NewEntryUnifiedApproach.LOOKUP_IDENTIFIER);
+
+        if (idText != null) {
+            Platform.runLater(() -> idText.requestFocus());
+        }
+
+        if (generateButton != null) {
+            generateButton.disableProperty().bind(idErrorInvalidText.visibleProperty().or(idErrorInvalidFetcher.visibleProperty()));
+            generateButton.setText("Lookup");
+        }
+    }
+
+    @FXML
+    private void switchInterpretCitations() {
+        if (!tabInterpretCitations.isSelected()) {
+            return;
+        }
+
+        currentApproach = NewEntryUnifiedApproach.INTERPRET_CITATIONS;
+        preferences.setLatestApproach(NewEntryUnifiedApproach.INTERPRET_CITATIONS);
+
+        if (interpretText != null) {
+            Platform.runLater(() -> interpretText.requestFocus());
+        }
+
+        if (generateButton != null) {
+            generateButton.disableProperty().bind(viewModel.interpretTextValidatorProperty().not());
+            generateButton.setText("Interpret");
+        }
+    }
+
+    @FXML
+    private void switchSpecifyBibtex() {
+        if (!tabSpecifyBibtex.isSelected()) {
+            return;
+        }
+
+        currentApproach = NewEntryUnifiedApproach.SPECIFY_BIBTEX;
+        preferences.setLatestApproach(NewEntryUnifiedApproach.SPECIFY_BIBTEX);
+
+        if (bibtexText != null) {
+            Platform.runLater(() -> bibtexText.requestFocus());
+        }
+
+        if (generateButton != null) {
+            generateButton.disableProperty().bind(viewModel.bibtexTextValidatorProperty().not());
+            generateButton.setText("Create");
+        }
+    }
+
+    private void onEntryTypeSelected(EntryType type) {
+        preferences.setLatestInstantType(type);
+        result = new BibEntry(type);
+        this.close();
+    }
+
+    private void onSuccessfulExecution() {
+        viewModel.cancel();
+        stateManager.activeSearchQuery(SearchType.NORMAL_SEARCH).set(Optional.empty());
+        this.close();
+    }
+
+    private void execute() {
+        // :TODO: These button text changes aren't actually visible, due to the UI thread not being able to perform the
+        // update before the button text is reset. The `viewModel.execute*()` and `switch*()` calls could be wrapped in
+        // a `Platform.runLater(...)` which would probably fix this.
+        switch (currentApproach) {
+            case NewEntryUnifiedApproach.CREATE_ENTRY:
+                // We do nothing here.
+                break;
+            case NewEntryUnifiedApproach.LOOKUP_IDENTIFIER:
+                generateButton.setText("Searching...");
+                viewModel.executeLookupIdentifier(idLookupGuess.isSelected());
+                switchLookupIdentifier();
+                break;
+            case NewEntryUnifiedApproach.INTERPRET_CITATIONS:
+                generateButton.setText("Parsing...");
+                viewModel.executeInterpretCitations();
+                switchInterpretCitations();
+                break;
+            case NewEntryUnifiedApproach.SPECIFY_BIBTEX:
+                generateButton.setText("Parsing...");
+                viewModel.executeSpecifyBibtex();
+                switchSpecifyBibtex();
+                break;
+        }
+    }
+
+    private void cancel() {
+        viewModel.cancel();
+    }
+
+    private void addEntriesToPane(FlowPane pane, Collection<? extends BibEntryType> entries) {
+        final double maxTooltipWidth = (2.0 / 3.0) * Screen.getPrimary().getBounds().getWidth();
+        for (BibEntryType entry : entries) {
+            final EntryType type = entry.getType();
+
+            final Button button = new Button(type.getDisplayName());
+            button.setUserData(entry);
+            button.setOnAction(event -> onEntryTypeSelected(type));
+            pane.getChildren().add(button);
+
+            final String description = descriptionOfEntryType(type);
+            if (description != null) {
+                final Tooltip tooltip = new Tooltip(description);
+                tooltip.setMaxWidth(maxTooltipWidth);
+                tooltip.setWrapText(true);
+                button.setTooltip(tooltip);
+            }
+        }
+    }
+
+    private static String descriptionOfEntryType(EntryType type) {
+        if (type instanceof StandardEntryType entryType) {
+            return descriptionOfStandardEntryType(entryType);
+        }
+        return null;
+    }
+
+    private static String descriptionOfStandardEntryType(StandardEntryType type) {
+        // These descriptions are taken from subsection 2.1 of the biblatex package documentation.
+        // Biblatex is a superset of bibtex, with more elaborate descriptions, so its documentation is preferred.
+        // See [https://mirrors.ibiblio.org/pub/mirrors/CTAN/macros/latex/contrib/biblatex/doc/biblatex.pdf].
+        return switch (type) {
+            case Article -> Localization.lang("An article in a journal, magazine, newspaper, or other periodical which forms a self-contained unit with its own title.");
+            case Book -> Localization.lang("A single-volume book with one or more authors where the authors share credit for the work as a whole.");
+            case Booklet -> Localization.lang("A book-like work without a formal publisher or sponsoring institution.");
+            case Collection -> Localization.lang("A single-volume collection with multiple, self-contained contributions by distinct authors which have their own title. The work as a whole has no overall author but it will usually have an editor.");
+            case Conference -> Localization.lang("A legacy alias for \"InProceedings\".");
+            case InBook -> Localization.lang("A part of a book which forms a self-contained unit with its own title.");
+            case InCollection -> Localization.lang("A contribution to a collection which forms a self-contained unit with a distinct author and title.");
+            case InProceedings -> Localization.lang("An article in a conference proceedings.");
+            case Manual -> Localization.lang("Technical or other documentation, not necessarily in printed form.");
+            case MastersThesis -> Localization.lang("Similar to \"Thesis\" except that the type field is optional and defaults to the localised term  Master's thesis.");
+            case Misc -> Localization.lang("A fallback type for entries which do not fit into any other category.");
+            case PhdThesis -> Localization.lang("Similar to \"Thesis\" except that the type field is optional and defaults to the localised term PhD thesis.");
+            case Proceedings -> Localization.lang("A single-volume conference proceedings. This type is very similar to \"Collection\".");
+            case TechReport -> Localization.lang("Similar to \"Report\" except that the type field is optional and defaults to the localised term technical report.");
+            case Unpublished -> Localization.lang("A work with an author and a title which has not been formally published, such as a manuscript or the script of a talk.");
+            case BookInBook -> Localization.lang("This type is similar to \"InBook\" but intended for works originally published as a stand-alone book.");
+            case InReference -> Localization.lang("An article in a work of reference. This is a more specific variant of the generic \"InCollection\" entry type.");
+            case MvBook -> Localization.lang("A multi-volume \"Book\".");
+            case MvCollection -> Localization.lang("A multi-volume \"Collection\".");
+            case MvProceedings -> Localization.lang("A multi-volume \"Proceedings\" entry.");
+            case MvReference -> Localization.lang("A multi-volume \"Reference\" entry. The standard styles will treat this entry type as an alias for \"MvCollection\".");
+            case Online -> Localization.lang("This entry type is intended for sources such as web sites which are intrinsically online resources.");
+            case Reference -> Localization.lang("A single-volume work of reference such as an encyclopedia or a dictionary.");
+            case Report -> Localization.lang("A technical report, research report, or white paper published by a university or some other institution.");
+            case Set -> Localization.lang("An entry set is a group of entries which are cited as a single reference and listed as a single item in the bibliography.");
+            case SuppBook -> Localization.lang("Supplemental material in a \"Book\". This type is provided for elements such as prefaces, introductions, forewords, afterwords, etc. which often have a generic title only.");
+            case SuppCollection -> Localization.lang("Supplemental material in a \"Collection\".");
+            case SuppPeriodical -> Localization.lang("Supplemental material in a \"Periodical\". This type may be useful when referring to items such as regular columns, obituaries, letters to the editor, etc. which only have a generic title.");
+            case Thesis -> Localization.lang("A thesis written for an educational institution to satisfy the requirements for a degree.");
+            case WWW -> Localization.lang("An alias for \"Online\", provided for jurabib compatibility.");
+            case Software -> Localization.lang("Computer software. The standard styles will treat this entry type as an alias for \"Misc\".");
+            case Dataset -> Localization.lang("A data set or a similar collection of (mostly) raw data.");
+        };
+    }
+
+    private static IdBasedFetcher fetcherFromName(String fetcherName, List<IdBasedFetcher> fetchers) {
+        for (IdBasedFetcher fetcher : fetchers) {
+            if (fetcher.getName().equals(fetcherName)) {
+                return fetcher;
+            }
+        }
+        return null;
+    }
+
+    private static PlainCitationParserChoice parserFromName(String parserName, List<PlainCitationParserChoice> parsers) {
+        for (PlainCitationParserChoice parser : parsers) {
+            if (parser.getLocalizedName().equals(parserName)) {
+                return parser;
+            }
+        }
+        return null;
+    }
+}