+++ jabgui/src/main/java/org/jabref/gui/preferences/GuiPreferences.java
@@ -14,6 +14,7 @@ import org.jabref.gui.maintable.ColumnPreferences;
 import org.jabref.gui.maintable.MainTablePreferences;
 import org.jabref.gui.maintable.NameDisplayPreferences;
 import org.jabref.gui.mergeentries.MergeDialogPreferences;
+import org.jabref.gui.newentryunified.NewEntryUnifiedPreferences;
 import org.jabref.gui.preview.PreviewPreferences;
 import org.jabref.gui.push.PushToApplicationPreferences;
 import org.jabref.gui.specialfields.SpecialFieldsPreferences;
@@ -55,4 +56,6 @@ public interface GuiPreferences extends CliPreferences {
     ColumnPreferences getSearchDialogColumnPreferences();
 
     KeyBindingRepository getKeyBindingRepository();
+
+    NewEntryUnifiedPreferences getNewEntryUnifiedPreferences();
 }