+++ jabgui/src/main/java/org/jabref/gui/preferences/JabRefGuiPreferences.java
@@ -3,6 +3,7 @@ package org.jabref.gui.preferences;
 import java.io.File;
 import java.nio.file.Path;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
@@ -40,6 +41,8 @@ import org.jabref.gui.maintable.MainTablePreferences;
 import org.jabref.gui.maintable.NameDisplayPreferences;
 import org.jabref.gui.mergeentries.DiffMode;
 import org.jabref.gui.mergeentries.MergeDialogPreferences;
+import org.jabref.gui.newentryunified.NewEntryUnifiedApproach;
+import org.jabref.gui.newentryunified.NewEntryUnifiedPreferences;
 import org.jabref.gui.preview.PreviewPreferences;
 import org.jabref.gui.push.PushToApplicationPreferences;
 import org.jabref.gui.push.PushToApplications;
@@ -55,6 +58,7 @@ import org.jabref.logic.exporter.SelfContainedSaveConfiguration;
 import org.jabref.logic.externalfiles.DateRange;
 import org.jabref.logic.externalfiles.ExternalFileSorter;
 import org.jabref.logic.importer.fetcher.DoiFetcher;
+import org.jabref.logic.importer.plaincitation.PlainCitationParserChoice;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.layout.TextBasedPreviewLayout;
@@ -67,6 +71,8 @@ import org.jabref.logic.util.StandardFileType;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.entry.field.Field;
 import org.jabref.model.entry.field.FieldFactory;
+import org.jabref.model.entry.types.EntryType;
+import org.jabref.model.entry.types.StandardEntryType;
 import org.jabref.model.groups.GroupHierarchyType;
 import org.jabref.model.metadata.SaveOrder;
 import org.jabref.model.metadata.SelfContainedSaveOrder;
@@ -210,7 +216,6 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
     private static final String SPECIALFIELDSENABLED = "specialFieldsEnabled";
     // endregion
 
-    private static final String ID_ENTRY_GENERATOR = "idEntryGenerator";
     private static final String SELECTED_SLR_CATALOGS = "selectedSlrCatalogs";
     private static final String UNLINKED_FILES_SELECTED_EXTENSION = "unlinkedFilesSelectedExtension";
     private static final String UNLINKED_FILES_SELECTED_DATE_RANGE = "unlinkedFilesSelectedDateRange";
@@ -218,6 +223,14 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
 
     private static final String INCLUDE_CROSS_REFERENCES = "includeCrossReferences";
     private static final String ASK_FOR_INCLUDING_CROSS_REFERENCES = "askForIncludingCrossReferences";
+
+    // region NewEntryUnifiedPreferences
+    private static final String NEU_APPROACH = "latestApproach";
+    private static final String NEU_INSTANT_TYPE = "latestInstantType";
+    private static final String NEU_ID_LOOKUP_GUESSING = "idLookupGuessing";
+    private static final String NEU_ID_FETCHER_NAME = "latestIdFetcherName";
+    private static final String NEU_INTERPRET_PARSER_NAME = "latestInterpretParserName";
+    // endregion
 
     private static JabRefGuiPreferences singleton;
 
@@ -239,6 +252,7 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
     private ColumnPreferences searchDialogColumnPreferences;
     private KeyBindingRepository keyBindingRepository;
     private CopyToPreferences copyToPreferences;
+    private NewEntryUnifiedPreferences newEntryUnifiedPreferences;
 
     private JabRefGuiPreferences() {
         super();
@@ -393,7 +407,6 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
 
         // region core GUI preferences
         // Set DOI to be the default ID entry generator
-        defaults.put(ID_ENTRY_GENERATOR, DoiFetcher.NAME);
         defaults.put(MAIN_WINDOW_POS_X, 0);
         defaults.put(MAIN_WINDOW_POS_Y, 0);
         defaults.put(MAIN_WINDOW_WIDTH, 1024);
@@ -405,6 +418,14 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
 
         defaults.put(ASK_FOR_INCLUDING_CROSS_REFERENCES, Boolean.TRUE);
         defaults.put(INCLUDE_CROSS_REFERENCES, Boolean.FALSE);
+
+        // region NewEntryUnifierPreferences
+        defaults.put(NEU_APPROACH, Arrays.asList(NewEntryUnifiedApproach.values()).indexOf(NewEntryUnifiedApproach.CREATE_ENTRY));
+        defaults.put(NEU_INSTANT_TYPE, StandardEntryType.Article.getDisplayName());
+        defaults.put(NEU_ID_LOOKUP_GUESSING, true);
+        defaults.put(NEU_ID_FETCHER_NAME, DoiFetcher.NAME);
+        defaults.put(NEU_INTERPRET_PARSER_NAME, PlainCitationParserChoice.RULE_BASED.getLocalizedName());
+        // endregion
     }
 
     /**
@@ -623,7 +644,6 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
                 getDouble(MAIN_WINDOW_WIDTH),
                 getDouble(MAIN_WINDOW_HEIGHT),
                 getBoolean(WINDOW_MAXIMISED),
-                get(ID_ENTRY_GENERATOR),
                 getDouble(SIDE_PANE_WIDTH),
                 getDouble(ENTRY_EDITOR_HEIGHT));
 
@@ -632,7 +652,6 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
         EasyBind.listen(coreGuiPreferences.sizeXProperty(), (_, _, newValue) -> putDouble(MAIN_WINDOW_WIDTH, newValue.doubleValue()));
         EasyBind.listen(coreGuiPreferences.sizeYProperty(), (_, _, newValue) -> putDouble(MAIN_WINDOW_HEIGHT, newValue.doubleValue()));
         EasyBind.listen(coreGuiPreferences.windowMaximisedProperty(), (_, _, newValue) -> putBoolean(WINDOW_MAXIMISED, newValue));
-        EasyBind.listen(coreGuiPreferences.lastSelectedIdBasedFetcherProperty(), (_, _, newValue) -> put(ID_ENTRY_GENERATOR, newValue));
         EasyBind.listen(coreGuiPreferences.horizontalDividerPositionProperty(), (_, _, newValue) -> putDouble(SIDE_PANE_WIDTH, newValue.doubleValue()));
         EasyBind.listen(coreGuiPreferences.getVerticalDividerPositionProperty(), (_, _, newValue) -> putDouble(ENTRY_EDITOR_HEIGHT, newValue.doubleValue()));
 
@@ -1230,6 +1249,42 @@ public class JabRefGuiPreferences extends JabRefCliPreferences implements GuiPre
         });
 
         return keyBindingRepository;
+    }
+
+    @Override
+    public NewEntryUnifiedPreferences getNewEntryUnifiedPreferences() {
+        if (newEntryUnifiedPreferences != null) {
+            return newEntryUnifiedPreferences;
+        }
+
+        final int approachIndex = getInt(NEU_APPROACH);
+        NewEntryUnifiedApproach approach = NewEntryUnifiedApproach.values().length > approachIndex
+            ? NewEntryUnifiedApproach.values()[approachIndex]
+            : NewEntryUnifiedApproach.values()[0];
+
+        final String instantTypeName = get(NEU_INSTANT_TYPE);
+        EntryType instantType = StandardEntryType.Article;
+        for (StandardEntryType type : StandardEntryType.values()) {
+            if (type.getDisplayName().equals(instantTypeName)) {
+                instantType = type;
+                break;
+            }
+        }
+
+        newEntryUnifiedPreferences = new NewEntryUnifiedPreferences(
+            approach,
+            instantType,
+            getBoolean(NEU_ID_LOOKUP_GUESSING),
+            get(NEU_ID_FETCHER_NAME),
+            get(NEU_INTERPRET_PARSER_NAME));
+
+        EasyBind.listen(newEntryUnifiedPreferences.latestApproachProperty(), (obs, oldValue, newValue) -> putInt(NEU_APPROACH, Arrays.asList(NewEntryUnifiedApproach.values()).indexOf(newValue)));
+        EasyBind.listen(newEntryUnifiedPreferences.latestInstantTypeProperty(), (obs, oldValue, newValue) -> put(NEU_INSTANT_TYPE, newValue.getDisplayName()));
+        EasyBind.listen(newEntryUnifiedPreferences.idLookupGuessingProperty(), (obs, oldValue, newValue) -> putBoolean(NEU_ID_LOOKUP_GUESSING, newValue));
+        EasyBind.listen(newEntryUnifiedPreferences.latestIdFetcherProperty(), (obs, oldValue, newValue) -> put(NEU_ID_FETCHER_NAME, newValue));
+        EasyBind.listen(newEntryUnifiedPreferences.latestInterpretParserProperty(), (obs, oldValue, newValue) -> put(NEU_INTERPRET_PARSER_NAME, newValue));
+
+        return newEntryUnifiedPreferences;
     }
 
     /**